+++ src/Ema/Asset.hs
@@ -6,6 +6,9 @@ module Ema.Asset where
 -- | The type of assets that can be bundled in a static site.
 data Asset a
   = -- | A file that is copied as-is from the source directory.
+    --
+    -- Relative paths are assumed relative to the source directory. Absolute
+    -- paths allow copying static files outside of source directory.
     AssetStatic FilePath
   | -- | A file whose contents are generated at runtime by user code.
     AssetGenerated Format a